🐛(serde,e2e): fix u64 JS overflow in wasm bindgen and add onboarding e2e test
- Change known_vocab_hash from u64 to u32: DefaultHasher values exceed
  JavaScript safe integer (2^53) causing serde_wasm_bindgen serialization
  failure in IndexedDB. u32 is sufficient for change detection.
- Add custom serde adapter for telegram_user_id (Option<u64>) to serialize
  as string for JS compatibility while deserializing both string and number
- Add e2e test verifying phrase cards appear after full onboarding with
  N5 level and Mark All Known scoring

+++ origa/src/use_cases/seed_ready_phrases.rs
@@ -93,14 +93,14 @@ fn collect_known_vocabulary_words(study_cards: &HashMap<Ulid, StudyCard>) -> Has
         .collect()
 }
 
-fn compute_known_words_hash(known_words: &HashSet<String>) -> u64 {
+fn compute_known_words_hash(known_words: &HashSet<String>) -> u32 {
     let mut words: Vec<&String> = known_words.iter().collect();
     words.sort();
     let mut hasher = std::collections::hash_map::DefaultHasher::new();
     for word in &words {
         word.hash(&mut hasher);
     }
-    hasher.finish()
+    hasher.finish() as u32
 }
 
 fn collect_existing_phrase_ids(study_cards: &HashMap<Ulid, StudyCard>) -> HashSet<Ulid> {
